refactor(#1483): share HTML link scanning for feed discovery
Extracts WebmentionEndpointDiscovery's generic <link>/<a> tag-scan and
attribute-parse logic into HTMLLinkAttributeScanning so a later task's
FeedEndpointDiscovery (RSS/Atom feed discovery for the blogroll's OPML
export) can reuse it instead of duplicating the same subtle regex logic.
Behavior-preserving: webmention-specific rel matching stays in place,
and all 13 existing WebmentionEndpointDiscovery tests pass unmodified.
