+++ Sources/AnglesiteCore/WebmentionEndpointDiscovery.swift
@@ -84,29 +84,8 @@ public enum WebmentionEndpointDiscovery {
 
     // MARK: HTML
 
-    /// Matches `<link ...>` and `<a ...>` tags in document order; the first with a `webmention`
-    /// rel wins, per the spec ("the first link or a element ... in document order").
-    ///
-    /// Known, accepted limitation: `[^>]*` truncates the tag at the first literal `>`, including
-    /// one embedded inside a quoted attribute value (e.g. `href="/x?a=1>2"`). A correct HTML
-    /// tokenizer would track quote state to know that `>` isn't a tag terminator there. This is a
-    /// conscious won't-fix, not an oversight — a literal, unencoded `>` inside an attribute value
-    /// is invalid per the URL spec (it must be percent-encoded as `%3E`) and vanishingly rare in
-    /// real-world markup; handling it would mean replacing this regex scan with a full tokenizer
-    /// for a case that essentially never occurs.
-    private static let tagPattern: NSRegularExpression = {
-        do {
-            return try NSRegularExpression(pattern: #"<(?:link|a)\b([^>]*)>"#, options: [.caseInsensitive])
-        } catch {
-            fatalError("Invalid webmention discovery tag regex: \(error)")
-        }
-    }()
-
     static func endpoint(fromHTML html: String, relativeTo baseURL: URL) -> URL? {
-        let range = NSRange(html.startIndex..<html.endIndex, in: html)
-        for match in tagPattern.matches(in: html, range: range) {
-            guard let attrsRange = Range(match.range(at: 1), in: html) else { continue }
-            let attrs = String(html[attrsRange])
+        for attrs in HTMLLinkAttributeScanning.tagAttributeStrings(in: html) {
             guard let rel = attributeValue("rel", in: attrs), isWebmentionRel(rel) else { continue }
             guard let href = attributeValue("href", in: attrs) else { continue }
             if let url = URL(string: href, relativeTo: baseURL)?.absoluteURL {
@@ -130,42 +109,7 @@ public enum WebmentionEndpointDiscovery {
         }
     }
 
-    /// Extracts `name="value"` / `name='value'` / `name=value` from an HTML tag's attribute
-    /// string or an HTTP Link-header parameter string. The lookahead-free anchor
-    /// `(?:^|[\s;<])` before `name` requires the name to start at the beginning of the source,
-    /// or be preceded by whitespace, a `;` (Link-header parameter separator), or `<` — so a
-    /// lookup for `rel` does not match inside a longer attribute name like `data-rel=`. (A
-    /// plain `\b` word-boundary anchor does *not* achieve this: `-` is a non-word character, so
-    /// `\brel\b` still matches the `rel` inside `data-rel=`.)
-    private static let relRegex = attributeRegex(for: "rel")
-    private static let hrefRegex = attributeRegex(for: "href")
-
-    private static func attributeRegex(for name: String) -> NSRegularExpression {
-        do {
-            return try NSRegularExpression(
-                pattern: "(?:^|[\\s;<])\(name)\\s*=\\s*(\"([^\"]*)\"|'([^']*)'|([^\\s\"'>]+))",
-                options: [.caseInsensitive]
-            )
-        } catch {
-            fatalError("Invalid webmention discovery attribute regex for \(name): \(error)")
-        }
-    }
-
     private static func attributeValue(_ name: String, in source: String) -> String? {
-        let regex: NSRegularExpression
-        switch name {
-        case "rel": regex = relRegex
-        case "href": regex = hrefRegex
-        default: regex = attributeRegex(for: name)
-        }
-        let range = NSRange(source.startIndex..<source.endIndex, in: source)
-        guard let match = regex.firstMatch(in: source, range: range) else { return nil }
-        for groupIndex in [2, 3, 4] {
-            let group = match.range(at: groupIndex)
-            if group.location != NSNotFound, let r = Range(group, in: source) {
-                return String(source[r])
-            }
-        }
-        return nil
+        HTMLLinkAttributeScanning.attributeValue(name, in: source)
     }
 }
